src/Lambda/requirements.txt
boto3==1.26.87
psycopg2-binary==2.9.5
python_dateutil==2.8.2
dateparser